refactor: create teamInfo file and extract constants inside
